Added modal on role removal on role details page

+++ src/app/pages/roles/role-has-users/role-has-users.ts
@@ -1,17 +1,19 @@
-import {Component, inject, signal, WritableSignal} from '@angular/core';
+import {Component, inject, signal, ViewChild, WritableSignal} from '@angular/core';
 import {RoleService} from '../../../services/role-service';
 import {ActivatedRoute, Router} from '@angular/router';
 import {User} from '../../../models/Users';
 import {Table} from '../../../components/table/table';
 import {UserWithRoleDto} from '../../../models/Role';
+import {Modal} from "../../../components/modal/modal";
 
 @Component({
   selector: 'app-role-has-users',
   imports: [
-    Table
+    Table,
+    Modal
   ],
   templateUrl: './role-has-users.html',
-  styleUrl: './role-has-users.css'
+  styleUrl: './role-has-users.scss'
 })
 export class RoleHasUsers {
   roleService = inject(RoleService);
@@ -20,6 +22,9 @@ export class RoleHasUsers {
 
   users: WritableSignal<Array<User>> = signal([]);
   role!: string;
+  userToRemove: WritableSignal<string | null> = signal(null);
+
+  @ViewChild('confirmDelete') confirmDeleteModal!: Modal;
 
 
   constructor() {
@@ -40,22 +45,41 @@ export class RoleHasUsers {
     })
   }
 
-  removeRole(username: string) {
+  openConfirmationModal(username: string) {
+    this.userToRemove.set(username);
+    this.confirmDeleteModal.title.set('Confirm Removal');
+    this.confirmDeleteModal.message.set(`Are you sure you want to remove the role '${this.role}' from user '${username}'?`);
+    this.confirmDeleteModal.open();
+  }
+
+  confirmRemove() {
+    const username = this.userToRemove();
+    if (!username) return;
+
     this.roleService.RemoveRoleFromUser(new UserWithRoleDto(this.role, username)).subscribe({
       next: (_) => {
         this.getUsers()
+        this.confirmDeleteModal.close();
+        this.userToRemove.set(null);
       },
       error: (err) => {
-        console.error('Error getting users:', err);
+        console.error('Error removing role from user:', err);
+        this.confirmDeleteModal.close();
+        this.userToRemove.set(null);
       }
     })
   }
 
-  back() {
-    this.router.navigate(['/Roles']).then(() => {});
+  cancelRemove() {
+    this.confirmDeleteModal.close();
+    this.userToRemove.set(null);
   }
 
-  details(username: string) {
-    this.router.navigate([`/Users/${username}`]).then(() => {});
+  async back() {
+    await this.router.navigate(['/Roles']);
+  }
+
+  async details(username: string) {
+    await this.router.navigate([`/Users/${username}`]);
   }
 }
